test: Split also plotting tests on their own
